Move f3d graph from lib to src
No git conflict after pull --rebase, but there was a semantic conflict
after restructuring. Fix by moving f3d to new folder structure
