+++ src/f3d/applyPatch.ts
